requirements.txt
django
django-widget-tweaks
djangorestframework
django-filter